backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-community
langchain-groq
langchain-huggingface

chromadb
pypdf

transformers
torch
sentence-transformers
